Rename table utils
