+++ src/components/SlateEditor/plugins/table/toolbarActions.ts
@@ -1,11 +1,18 @@
 import { Editor, Element, Path, Transforms } from 'slate';
 import { jsx as slatejsx } from 'slate-hyperscript';
 import { ReactEditor } from 'slate-react';
-import { TableElement, TableRowElement, TableHeadElement, TableBodyElement } from './interfaces';
+import { TableElement } from './interfaces';
 import getCurrentBlock from '../../utils/getCurrentBlock';
 import { getTableAsMatrix, getTableBodyAsMatrix } from './matrix';
 import { TYPE_TABLE_CELL, TYPE_TABLE_ROW, TYPE_TABLE_BODY } from './types';
-import { isTable, isTableBody, isTableCell, isTableHead, isTableRow } from './slateHelpers';
+import {
+  getTableBodyWidth,
+  isTable,
+  isTableBody,
+  isTableCell,
+  isTableHead,
+  isTableRow,
+} from './slateHelpers';
 import { findCellCoordinate } from './matrixHelpers';
 import { updateCell } from './slateActions';
 import {
@@ -14,49 +21,6 @@ import {
   defaultTableRowBlock,
   defaultTableBodyBlock,
 } from './defaultBlocks';
-
-export const countCells = (row: TableRowElement, stop?: number) => {
-  return row.children
-    .map(child => {
-      if (!isTableCell(child)) {
-        return 0;
-      }
-      return child.data.colspan;
-    })
-    .slice(0, stop)
-    .reduce((a, b) => a + b);
-};
-
-export const getTableBodyWidth = (element: TableHeadElement | TableBodyElement) => {
-  const firstRow = element.children[0];
-  if (isTableRow(firstRow)) {
-    return countCells(firstRow);
-  }
-  return 0;
-};
-
-export const getTableBodyHeight = (element: TableHeadElement | TableBodyElement) => {
-  return element.children.length;
-};
-
-export const createIdenticalRow = (element: TableRowElement) => {
-  return slatejsx(
-    'element',
-    { type: TYPE_TABLE_ROW },
-    element.children.map(child => {
-      if (isTableCell(child)) {
-        return {
-          ...defaultTableCellBlock(),
-          data: {
-            ...child.data,
-            rowspan: 1,
-          },
-        };
-      }
-      return defaultTableCellBlock();
-    }),
-  );
-};
 
 export const toggleRowHeaders = (editor: Editor, path: Path) => {
   const [table] = Editor.node(editor, path);
